djangoWords/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
environ==1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
